Change filename to useAuth.js and add documentation
